refactor: split up the WAL module
